service--payment, kafka publisher e consumer

+++ commons/src/main/java/io/github/thatyane/sagachoreography/commons/application/dto/OrderResponse.java
@@ -4,15 +4,13 @@ import io.github.thatyane.sagachoreography.commons.application.dto.enums.OrderSt
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 public class OrderResponse {
 
-    private Integer userId;
-    private Integer productId;
-    private BigDecimal amount;
+    private Long userId;
+    private Long productId;
+    private Integer amount;
     private Long orderId;
     private OrderStatus orderStatus;
 }
